change gpr_atm to int

+++ test/cpp/qps/server_async.cc
@@ -157,7 +157,7 @@ class AsyncQpsServerTest final : public grpc::testing::Server {
   int GetPollCount() override {
     int count = 0;
     for (auto cq = srv_cqs_.begin(); cq != srv_cqs_.end(); cq++) {
-      count += (int)grpc_get_cq_poll_num((*cq)->cq());
+      count += grpc_get_cq_poll_num((*cq)->cq());
     }
     return count;
   }
